Google Meet: Extract shared ProfileCommand to deduplicate manual-profile entries
manual-profile.tsx and manual-profile-and-refocus.tsx were verbatim
copies differing only in passing refocus to ProfileList. Extracted the
shared body into components/ProfileCommand.tsx parameterised by an
optional refocus prop, and reduced both command entry files to thin
wrappers. Prevents drift if the profile-picker UI changes later.
