+++ extensions/google-meet/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./ProfileList";
 export * from "./ProfileForm";
+export * from "./ProfileCommand";
